Combine GCSReadBenchmark and S3ReadBenchmark with ReadBenchmark
Move ReadBenchmark to benchmarks folder
