+++ velox/common/file/CMakeLists.txt
@@ -23,6 +23,3 @@ target_link_libraries(
 if(${VELOX_BUILD_TESTING})
   add_subdirectory(tests)
 endif()
-if(${VELOX_ENABLE_BENCHMARKS})
-  add_subdirectory(benchmark)
-endif()
