+++ velox/connectors/hive/storage_adapters/gcs/CMakeLists.txt
@@ -25,10 +25,6 @@ if(VELOX_ENABLE_GCS)
     add_subdirectory(tests)
   endif()
 
-  if(${VELOX_ENABLE_BENCHMARKS})
-    add_subdirectory(benchmark)
-  endif()
-
   if(${VELOX_ENABLE_EXAMPLES})
     add_subdirectory(examples)
   endif()
