+++ velox/connectors/hive/storage_adapters/s3fs/CMakeLists.txt
@@ -25,7 +25,4 @@ if(VELOX_ENABLE_S3)
   if(${VELOX_BUILD_TESTING})
     add_subdirectory(tests)
   endif()
-  if(${VELOX_ENABLE_BENCHMARKS})
-    add_subdirectory(benchmark)
-  endif()
 endif()
