--- velox/connectors/hive/storage_adapters/s3fs/benchmark/CMakeLists.txt
@@ -1,25 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-add_executable(velox_s3read_benchmark S3ReadBenchmark.cpp
-                                      S3ReadBenchmarkMain.cpp)
-
-target_link_libraries(
-  velox_s3read_benchmark
-  velox_read_benchmark_lib
-  velox_s3fs
-  velox_exception
-  velox_exec_test_lib
-  fmt::fmt
-  Folly::folly)
